+++ src/components/Main/Selector/Selector.styles.jsx
@@ -2,6 +2,10 @@ import { styled } from "styled-components";
 
 export const $Selector = styled.div`
   display: block;
+  
+  & > * {
+    margin-bottom: 15px;
+  }
 `;
 
 export default $Selector;