+++ src/store/selector-context.jsx
@@ -5,25 +5,50 @@ export const SelectorContext = createContext({
   setShowCriteria: () => {},
   whereCriteria: [],
   setWhereCriteria: () => {},
-  tempWhereCriteria: {key: "", value: ""},
+  clearWhereCriteria: () => {},
+  deleteWhereCriteria: () => {},
+  tempWhereCriteria: { key: "", value: "", id: "" },
   setTempWhereCriteria: () => {},
   selectorFilters: [],
-  setSelectorFilters: () => {}
+  setSelectorFilters: () => {},
+  calculatedMetrics: [],
+  setCalculatedMetrics: () => {}
 });
 
 function selectorReducer(state, action) {
-  if (action.type === 'SET_SHOW_CRITERIA') {
+  if (action.type === "SET_SHOW_CRITERIA") {
     return {
       ...state,
-      showCriteria: action.payload
-    }
+      showCriteria: action.payload,
+    };
   }
 
-  if (action.type === 'SET_WHERE_CRITERIA') {
-    const updatedWhereCriteria = [...state.whereCriteria];
+  if (action.type === "CLEAR_WHERE_CRITERIA") {
+    let updatedWhereCriteria = [...state.whereCriteria];
+    updatedWhereCriteria.length = 0;
 
-    const existingWhereCriteriaIndex = updatedWhereCriteria.findIndex(criteria => criteria.key === action.payload.key);
-    if (existingWhereCriteriaIndex) {
+    return {
+      ...state,
+      whereCriteria: [...updatedWhereCriteria],
+    };
+  }
+
+  if (action.type === "DELETE_WHERE_CRITERIA") {
+    const updatedWhereCriteria = [...state.whereCriteria];
+    updatedWhereCriteria.splice(action.payload, 1);
+
+    return {
+      ...state,
+      whereCriteria: [...updatedWhereCriteria],
+    };
+  }
+
+  if (action.type === "SET_WHERE_CRITERIA") {
+    const updatedWhereCriteria = [...state.whereCriteria];
+    const existingWhereCriteriaIndex = updatedWhereCriteria.findIndex(
+      (criteria) => criteria.key === action.payload.key
+    );
+    if (existingWhereCriteriaIndex !== -1) {
       updatedWhereCriteria[existingWhereCriteriaIndex] = action.payload;
     } else {
       updatedWhereCriteria.push(action.payload);
@@ -31,21 +56,28 @@ function selectorReducer(state, action) {
 
     return {
       ...state,
-      whereCriteria: updatedWhereCriteria
-    }
+      whereCriteria: [...updatedWhereCriteria],
+    };
   }
 
-  if (action.type === 'SET_TEMP_WHERE_CRITERIA') {
+  if (action.type === "SET_TEMP_WHERE_CRITERIA") {
     return {
       ...state,
-      tempWhereCriteria: action.payload
-    }
+      tempWhereCriteria: action.payload,
+    };
   }
 
-  if (action.type === 'SET_SELECTOR_FILTERS') {
+  if (action.type === "SET_SELECTOR_FILTERS") {
     return {
       ...state,
-      selectorFilters: action.payload
+      selectorFilters: action.payload,
+    };
+  }
+
+  if (action.type === "SET_CALCULATED_METRICS") {
+    return {
+      ...state,
+      calculatedMetrics: action.payload,
     }
   }
 
@@ -53,41 +85,59 @@ function selectorReducer(state, action) {
 }
 
 export default function SelectorContextProvider({ children }) {
-  const [selectorState, selectorDispatch] = useReducer(
-    selectorReducer,
-    {
-      showCriteria: "",
-      whereCriteria: [],
-      tempWhereCriteria: {key: "", value: ""},
-      selectorFilters: []
-    }
-  )
+  const [selectorState, selectorDispatch] = useReducer(selectorReducer, {
+    showCriteria: "",
+    whereCriteria: [],
+    tempWhereCriteria: { key: "", value: "" },
+    selectorFilters: [],
+    calculatedMetrics: [],
+  });
 
   function setShowCriteriaHandler(showCriteria) {
     selectorDispatch({
-      type: 'SET_SHOW_CRITERIA',
-      payload: showCriteria
-    })
+      type: "SET_SHOW_CRITERIA",
+      payload: showCriteria,
+    });
+  }
+
+  function clearWhereCriteriaHandler() {
+    selectorDispatch({
+      type: "CLEAR_WHERE_CRITERIA",
+    });
+  }
+
+  function deleteWhereCriteriaHandler(criteriaIndex) {
+    selectorDispatch({
+      type: "DELETE_WHERE_CRITERIA",
+      payload: criteriaIndex,
+    });
   }
 
   function setWhereCriteriaHandler(whereCriteria) {
     selectorDispatch({
-      type: 'SET_WHERE_CRITERIA',
-      payload: whereCriteria
-    })
+      type: "SET_WHERE_CRITERIA",
+      payload: whereCriteria,
+    });
   }
 
   function setTempWhereCriteriaHandler(tempWhereCriteria) {
     selectorDispatch({
-      type: 'SET_TEMP_WHERE_CRITERIA',
-      payload: tempWhereCriteria
-    })
+      type: "SET_TEMP_WHERE_CRITERIA",
+      payload: tempWhereCriteria,
+    });
   }
 
   function setSelectorFiltersHandler(selectorFilters) {
     selectorDispatch({
-      type: 'SET_SELECTOR_FILTERS',
-      payload: selectorFilters
+      type: "SET_SELECTOR_FILTERS",
+      payload: selectorFilters,
+    });
+  }
+
+  function setCalculatedMetricsHandler(calculatedMetrics) {
+    selectorDispatch({
+      type: "SET_CALCULATED_METRICS",
+      payload: calculatedMetrics,
     })
   }
 
@@ -96,15 +146,19 @@ export default function SelectorContextProvider({ children }) {
     setShowCriteria: setShowCriteriaHandler,
     whereCriteria: selectorState.whereCriteria,
     setWhereCriteria: setWhereCriteriaHandler,
+    clearWhereCriteria: clearWhereCriteriaHandler,
+    deleteWhereCriteria: deleteWhereCriteriaHandler,
     tempWhereCriteria: selectorState.tempWhereCriteria,
     setTempWhereCriteria: setTempWhereCriteriaHandler,
     selectorFilters: selectorState.selectorFilters,
-    setSelectorFilters: setSelectorFiltersHandler
-  }
+    setSelectorFilters: setSelectorFiltersHandler,
+    calculatedMetrics: selectorState.calculatedMetrics,
+    setCalculatedMetrics: setCalculatedMetricsHandler,
+  };
 
   return (
-    <SelectorContext.Provider value={ctxSelector}>{children}</SelectorContext.Provider>
+    <SelectorContext.Provider value={ctxSelector}>
+      {children}
+    </SelectorContext.Provider>
   );
 }
-
-