+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+0.22 (2024-08-04)
+===============
+* [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/12)
+* Create filters section and logic
+* Extend selector context by deleting filters
+* Implement id in filter value object
+* Create new api call to get calculated metrics and implement it into proper button
+* Implement charts for selector and proper functions to calculate usage and footprint
+
 0.21 (2024-07-27)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/11)